feat: migration loader class to typescript
